+++ jdplus-sdmx-desktop/jdplus-sdmx-desktop-plugin/src/main/java/jdplus/sdmx/desktop/plugin/web/SdmxWebConfiguration.java
@@ -12,19 +12,25 @@ import org.openide.nodes.Sheet;
 import sdmxdl.Languages;
 import sdmxdl.web.SdmxWebManager;
 import sdmxdl.web.WebSource;
+import standalone_sdmxdl.nbbrd.io.text.BooleanProperty;
 import standalone_sdmxdl.nbbrd.io.text.Parser;
+import standalone_sdmxdl.nbbrd.io.text.Property;
+import standalone_sdmxdl.sdmxdl.provider.PropertiesSupport;
 import standalone_sdmxdl.sdmxdl.provider.ri.caching.RiCaching;
-import standalone_sdmxdl.sdmxdl.provider.ri.drivers.SourceProperties;
 import standalone_sdmxdl.sdmxdl.provider.ri.networking.RiNetworking;
+import standalone_sdmxdl.sdmxdl.provider.ri.registry.RiRegistry;
 
 import javax.swing.filechooser.FileNameExtensionFilter;
 import java.io.File;
 import java.io.IOException;
-import java.util.Collections;
-import java.util.List;
 import java.util.Locale;
 import java.util.Properties;
+import java.util.function.Function;
+import java.util.logging.Level;
 
+import static java.util.Collections.emptyMap;
+
+@lombok.extern.java.Log
 @lombok.Data
 public class SdmxWebConfiguration {
 
@@ -60,6 +66,10 @@ public class SdmxWebConfiguration {
     private static final boolean DEFAULT_DISPLAY_CODES = false;
     private boolean displayCodes = DEFAULT_DISPLAY_CODES;
 
+    private static final String LOG_EVENTS_PROPERTY = "logEvents";
+    private static final boolean DEFAULT_LOG_EVENTS = false;
+    private boolean logEvents = DEFAULT_LOG_EVENTS;
+
     @MightBeGenerated
     public static SdmxWebConfiguration copyOf(SdmxWebConfiguration bean) {
         SdmxWebConfiguration result = new SdmxWebConfiguration();
@@ -71,31 +81,43 @@ public class SdmxWebConfiguration {
         result.noDefaultSSL = bean.noDefaultSSL;
         result.noSystemSSL = bean.noSystemSSL;
         result.displayCodes = bean.displayCodes;
+        result.logEvents = bean.logEvents;
         return result;
     }
 
     public SdmxWebManager toSdmxWebManager() {
         Properties properties = System.getProperties();
 
-        curlBackend.applyTo(properties, RiNetworking.CURL_BACKEND_PROPERTY);
+        if (sources != null && !sources.getPath().isEmpty()) {
+            properties.setProperty(RiRegistry.SOURCES_FILE_PROPERTY.getKey(), sources.getPath());
+        } else {
+            properties.remove(RiRegistry.SOURCES_FILE_PROPERTY.getKey());
+        }
+
+        curlBackend.applyTo(properties, RiNetworking.URL_BACKEND_PROPERTY, "JDK", "CURL");
         noCache.applyTo(properties, RiCaching.NO_CACHE_PROPERTY);
         autoProxy.applyTo(properties, RiNetworking.AUTO_PROXY_PROPERTY);
         noDefaultSSL.applyTo(properties, RiNetworking.NO_DEFAULT_SSL_PROPERTY);
         noSystemSSL.applyTo(properties, RiNetworking.NO_SYSTEM_SSL_PROPERTY);
 
+        logConfig();
+
         return SdmxWebManager.ofServiceLoader()
                 .toBuilder()
                 .onEvent(this::reportEvent)
                 .onError(this::reportError)
-                .customSources(getCustomSources())
                 .build();
     }
 
-    private static List<WebSource> getCustomSources() {
-        try {
-            return SourceProperties.loadCustomSources();
-        } catch (IOException e) {
-            return Collections.emptyList();
+    private void logConfig() {
+        Function<? super String, ? extends CharSequence> properties = key -> PropertiesSupport.getProperty(emptyMap(), key);
+        if (log.isLoggable(Level.INFO)) {
+            for (Property<?> p : new Property<?>[]{RiRegistry.SOURCES_FILE_PROPERTY, RiNetworking.URL_BACKEND_PROPERTY}) {
+                log.log(Level.INFO, p.getKey() + ": " + p.get(properties));
+            }
+            for (BooleanProperty p : new BooleanProperty[]{RiCaching.NO_CACHE_PROPERTY, RiNetworking.AUTO_PROXY_PROPERTY, RiNetworking.NO_DEFAULT_SSL_PROPERTY, RiNetworking.NO_SYSTEM_SSL_PROPERTY}) {
+                log.log(Level.INFO, p.getKey() + ": " + p.get(properties));
+            }
         }
     }
 
@@ -107,10 +129,18 @@ public class SdmxWebConfiguration {
 
     private void reportEvent(WebSource source, String marker, CharSequence message) {
         StatusDisplayer.getDefault().setStatusText(message.toString());
+        if (logEvents) {
+            log.log(Level.INFO, () -> asLogMessage(source, marker, message));
+        }
     }
 
     private void reportError(WebSource source, String marker, CharSequence message, IOException error) {
         NotificationDisplayer.getDefault().notify(message.toString(), SdmxIcons.getDefaultIcon(), "", null);
+        log.log(Level.SEVERE, error, () -> asLogMessage(source, marker, message));
+    }
+
+    private static String asLogMessage(WebSource source, String marker, CharSequence message) {
+        return "[" + source.getId() + "] (" + marker + ") " + message;
     }
 
     Sheet toSheet() {
@@ -159,6 +189,11 @@ public class SdmxWebConfiguration {
                 .display("No system SSL")
                 .description("Disable system truststore")
                 .add();
+        b.withBoolean()
+                .select(this, LOG_EVENTS_PROPERTY)
+                .display("Log events")
+                .description("Log events in IDE logs")
+                .add();
         result.put(b.build());
 
         return result;
@@ -177,5 +212,6 @@ public class SdmxWebConfiguration {
             .with(PropertyHandler.onEnum(NO_DEFAULT_SSL_PROPERTY, DEFAULT_NO_DEFAULT_SSL), SdmxWebConfiguration::getNoDefaultSSL, SdmxWebConfiguration::setNoDefaultSSL)
             .with(PropertyHandler.onEnum(NO_SYSTEM_SSL_PROPERTY, DEFAULT_NO_SYSTEM_SSL), SdmxWebConfiguration::getNoSystemSSL, SdmxWebConfiguration::setNoSystemSSL)
             .with(PropertyHandler.onBoolean(DISPLAY_CODES_PROPERTY, DEFAULT_DISPLAY_CODES), SdmxWebConfiguration::isDisplayCodes, SdmxWebConfiguration::setDisplayCodes)
+            .with(PropertyHandler.onBoolean(LOG_EVENTS_PROPERTY, DEFAULT_LOG_EVENTS), SdmxWebConfiguration::isLogEvents, SdmxWebConfiguration::setLogEvents)
             .build();
 }