+++ CHANGELOG.md
@@ -7,6 +7,23 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+## [3.2.0] - 2024-10-04
+
+This is a feature release of SDMX extensions for JDemetra+ v3.  
+[JDemetra+ v3.1.1 or later](https://github.com/jdemetra/jdplus-main) version is required to run it.
+
+### Added
+
+- Add logging of events
+
+### Fixed
+
+- Fix sources file property in configuration
+
+### Changed
+
+- Bump sdmx-dl from 3.0.0-beta.12 to [3.0.0-beta.13](https://github.com/nbbrd/sdmx-dl/releases/tag/v3.0.0-beta.13)
+
 ## [3.1.0] - 2023-10-16
 
 This is a feature release of SDMX extensions for JDemetra+ v3.  
@@ -30,6 +47,7 @@ This is a feature release of SDMX extensions for JDemetra+ v3.
 This is the **initial release** of SDMX extensions for JDemetra+ v3.  
 [JDemetra+ v3.0.2 or later](https://github.com/jdemetra/jdplus-main) version is required to run it.
 
-[Unreleased]: https://github.com/nbbrd/jdplus-sdmx/compare/v3.1.0...HEAD
+[Unreleased]: https://github.com/nbbrd/jdplus-sdmx/compare/v3.2.0...HEAD
+[3.2.0]: https://github.com/nbbrd/jdplus-sdmx/compare/v3.1.0...v3.2.0
 [3.1.0]: https://github.com/nbbrd/jdplus-sdmx/compare/v3.0.0...v3.1.0
 [3.0.0]: https://github.com/nbbrd/jdplus-sdmx/releases/tag/v3.0.0